requirements.txt
Flask==2.2.2
spotipy==2.21.0
lyricsgenius==3.0.1
nltk==3.8.1
vaderSentiment==3.3.2
wordcloud==1.8.2.2
pyenchant==3.2.2